+++ test-packages/custom-finger-print-app/.ember-cli
