+++ test-packages/custom-finger-print-app/.watchmanconfig
