+++ test-packages/custom-finger-print-app/README.md
@@ -1,4 +1,4 @@
-# basic-app
+# custom-app
 
 This README outlines the details of collaborating on this Ember application.
 A short introduction of this app could easily go here.
@@ -15,7 +15,7 @@ You will need the following things properly installed on your computer.
 ## Installation
 
 * `git clone <repository-url>` this repository
-* `cd basic-app`
+* `cd custom-app`
 * `npm install`
 
 ## Running / Development
