+++ test-packages/custom-finger-print-app/config/fastboot.js
