+++ test-packages/custom-finger-print-app/config/optional-features.json
