+++ test-packages/custom-finger-print-app/config/targets.js
