+++ test-packages/custom-finger-print-app/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "basic-app",
+  "name": "custom-app",
   "version": "0.0.0",
   "private": true,
-  "description": "Small description for basic-app goes here",
+  "description": "Small description for custom-app goes here",
   "repository": "",
   "license": "MIT",
   "author": "",
@@ -51,7 +51,7 @@
     "eslint": "^6.8.0",
     "eslint-plugin-ember": "^7.10.1",
     "eslint-plugin-node": "^11.0.0",
-    "fake-addon": "*",
+    "fake-addon-with-fastboot-config": "*",
     "fake-addon-2": "*",
     "loader.js": "^4.7.0",
     "mocha": "^8.0.1",
